Working Release of Event Master
All features are now up and running.

Now, users can make their attendance status social with all their other
friends on Omlet. As an added plus, users can also figure out who is
attending and who isn’t.

UI Updates as well. Now, the creator of the event can select a “theme”
for the page, which basically just changes the title bar banner image
to suit whatever event they are planning.

There are still some performance bugs, and we are working on it.
Additionally, there are some other features that we will be working to
bring to you all.

E